Updates AGWA_Delineate_Watershed.py
Refactors code to write delineation metadata before the delineation instead of after.
This results in the delineate function taking the delineation workspace and delineation name as parameters, which are used to query the metadata to obtain the associated delineation inputs instead of passing the inputs as parameters to the function.