Adds joining of streams layer to Join Results tool.
Current tool operation assumes both elements and streams layers have results joined to them.
The tool will fail if only one of the layers has the join.
It will not fail if the streams layer is joined to a different simulation than the elements layer, however this is not explicitly handled and should be.